automerge

+++ main/admin/usergroup_users.php
@@ -1,10 +1,6 @@
 <?php
 /* For licensing terms, see /license.txt */
 
-/**
- *  @package chamilo.admin
- */
-$cidReset = true;
 require_once __DIR__.'/../inc/global.inc.php';
 
 $this_section = SECTION_PLATFORM_ADMIN;
@@ -13,15 +9,15 @@ $id = isset($_GET['id']) ? (int) $_GET['id'] : 0;
 
 $usergroup = new UserGroup();
 $userGroupInfo = $usergroup->get($id);
-
 if (empty($userGroupInfo)) {
     api_not_allowed(true);
 }
 
-$usergroup->protectScript($userGroupInfo);
+$usergroup->protectScript($userGroupInfo, true, true);
+$allowEdit = api_is_platform_admin() || isset($userGroupInfo['author_id']) && $userGroupInfo['author_id'] == api_get_user_id();
 
 $calendarPlugin = null;
-if (api_get_plugin_setting('learning_calendar', 'enabled') === 'true') {
+if ($allowEdit && api_get_plugin_setting('learning_calendar', 'enabled') === 'true') {
     $calendarPlugin = LearningCalendarPlugin::create();
 }
 
@@ -31,9 +27,17 @@ $action = isset($_GET['action']) ? Security::remove_XSS($_GET['action']) : null;
 $userId = isset($_GET['user_id']) ? (int) $_GET['user_id'] : 0;
 $calendarId = isset($_REQUEST['calendar_id']) ? (int) $_REQUEST['calendar_id'] : 0;
 
-// setting breadcrumbs
-$interbreadcrumb[] = ['url' => 'usergroups.php', 'name' => get_lang('Classes')];
+$courseInfo = api_get_course_info();
+if (empty($courseInfo)) {
+    $interbreadcrumb[] = ['url' => 'usergroups.php', 'name' => get_lang('Classes')];
+} else {
+    $interbreadcrumb[] = ['url' => api_get_path(WEB_CODE_PATH).'user/class.php?'.api_get_cidreq(), 'name' => get_lang('Classes')];
+}
 $interbreadcrumb[] = ['url' => '#', 'name' => $userGroupInfo['name']];
+
+if (!empty($action)) {
+    $usergroup->protectScript($userGroupInfo);
+}
 
 switch ($action) {
     case 'add_calendar':
@@ -94,7 +98,7 @@ switch ($action) {
         $value = isset($_GET['value']) ? (int) $_GET['value'] : 0;
         $calendarPlugin->addControlPoint($userId, $value);
         Display::addFlash(
-            Display::return_message($calendarPlugin->get_lang('ControlPointAdded'), 'confirmation')
+            Display::return_message($calendarPlugin->get_lang('Control point added'), 'confirmation')
         );
         header('Location: '.api_get_self().'?id='.$id);
         exit;
@@ -124,7 +128,7 @@ $url = api_get_path(WEB_AJAX_PATH).'model.ajax.php?a=get_usergroups_users&id='.$
 // The order is important you need to check the the $column variable in the model.ajax.php file
 $columns = [
     get_lang('Name'),
-    get_lang('Actions'),
+    get_lang('Detail'),
 ];
 
 // Column config
@@ -140,15 +144,15 @@ $column_model = [
     ],
 ];
 
-if (api_get_plugin_setting('learning_calendar', 'enabled') === 'true') {
+if ($calendarPlugin) {
     $columns = [
         get_lang('Name'),
         get_lang('Calendar'),
-        get_lang('ClassroomActivity'),
-        get_lang('TimeSpentByStudentsInCourses'),
-        $calendarPlugin->get_lang('NumberDaysAccumulatedInCalendar'),
-        $calendarPlugin->get_lang('DifferenceOfDaysAndCalendar'),
-        get_lang('Actions'),
+        get_lang('Classroom activity'),
+        get_lang('Time spent by students in courses'),
+        $calendarPlugin->get_lang('Number of days accumulated in calendar'),
+        $calendarPlugin->get_lang('Difference between days and calendar'),
+        get_lang('Detail'),
     ];
 
     // Column config
@@ -195,25 +199,33 @@ $extraParams['autowidth'] = 'true';
 $extraParams['height'] = 'auto';
 $extraParams['sortname'] = 'name';
 $extraParams['sortorder'] = 'desc';
-$extraParams['multiselect'] = true;
+$extraParams['multiselect'] = $allowEdit;
 
 $deleteIcon = Display::return_icon('delete.png', get_lang('Delete'), null, ICON_SIZE_SMALL);
 $urlStats = api_get_path(WEB_CODE_PATH);
 
-//$addCalendar = '<a href="'.$urlStats.'mySpace/myStudents.php?student=\'+options.rowId+\'">'.Display::return_icon('agenda.png', get_lang('Agenda'), '', ICON_SIZE_SMALL).'</a>';
-
 $reportingIcon = Display::return_icon('statistics.png', get_lang('Reporting'), '', ICON_SIZE_SMALL);
-$controlPoint = Display::return_icon('add.png', get_lang('ControlPoint'), '', ICON_SIZE_SMALL);
+$controlPoint = Display::return_icon('add.png', get_lang('Control point'), '', ICON_SIZE_SMALL);
 
+$link = '';
+
+if ($calendarPlugin) {
+    $link = '<a href="'.$urlStats.'admin/usergroup_users.php?action=create_control_point&value=\'+value+\'&id='.$id.'&user_id=\'+options.rowId+\'">'.$controlPoint.'</a>';
+}
+
+$deleteButton = '';
+if ($allowEdit) {
+    $deleteButton = '<a onclick="javascript:if(!confirm('."\'".addslashes(api_htmlentities(get_lang('Please confirm your choice'), ENT_QUOTES))."\'".')) return false;" href="?id='.$id.'&action=delete&user_id=\'+options.rowId+\'">'.$deleteIcon.'</a>';
+}
 //return \'<a href="session_edit.php?page=resume_session.php&id=\'+options.rowId+\'">'.Display::return_icon('edit.png', get_lang('Edit'), '', ICON_SIZE_SMALL).'</a>'.
 // With this function we can add actions to the jgrid
 $action_links = '
 function action_formatter(cellvalue, options, rowObject) {
     var value = rowObject[5];
     return \''.
-    '&nbsp;<a href="'.$urlStats.'admin/usergroup_users.php?action=create_control_point&value=\'+value+\'&id='.$id.'&user_id=\'+options.rowId+\'">'.$controlPoint.'</a>'.
+    '&nbsp;'.$link.
     '&nbsp;<a href="'.$urlStats.'mySpace/myStudents.php?student=\'+options.rowId+\'">'.$reportingIcon.'</a>'.
-    ' <a onclick="javascript:if(!confirm('."\'".addslashes(api_htmlentities(get_lang('ConfirmYourChoice'), ENT_QUOTES))."\'".')) return false;" href="?id='.$id.'&action=delete&user_id=\'+options.rowId+\'">'.$deleteIcon.'</a>\';
+    ' '.$deleteButton.' \';
 }
 
 function extra_formatter(cellvalue, options, rowObject) {
@@ -232,47 +244,50 @@ function extra_formatter(cellvalue, options, rowObject) {
 
     return \''.
     '&nbsp;<a href="'.$urlStats.'mySpace/myStudents.php?student=\'+options.rowId+\'">'.Display::return_icon('statistics.png', get_lang('Reporting'), '', ICON_SIZE_SMALL).'</a>'.
-    ' <a onclick="javascript:if(!confirm('."\'".addslashes(api_htmlentities(get_lang("ConfirmYourChoice"), ENT_QUOTES))."\'".')) return false;"  href="?id='.$id.'&action=delete&user_id=\'+options.rowId+\'">'.$deleteIcon.'</a>\';
+    ' <a onclick="javascript:if(!confirm('."\'".addslashes(api_htmlentities(get_lang("Please confirm your choice"), ENT_QUOTES))."\'".')) return false;"  href="?id='.$id.'&action=delete&user_id=\'+options.rowId+\'">'.$deleteIcon.'</a>\';
 }';
 
 $deleteUrl = api_get_path(WEB_AJAX_PATH).'usergroup.ajax.php?a=delete_user_in_usergroup&group_id='.$id;
 
-$form = new FormValidator(
-    'add_multiple_calendar',
-    'post',
-    api_get_self().'?id='.$id.'&action=add_multiple_users_to_calendar'
-);
-$calendarPlugin->getAddUserToCalendarForm($form);
-$form->addHidden('user_list', '');
-$form->addButtonSave(get_lang('Add'));
+if ($calendarPlugin) {
+    $form = new FormValidator(
+        'add_multiple_calendar',
+        'post',
+        api_get_self().'?id='.$id.'&action=add_multiple_users_to_calendar'
+    );
+    $calendarPlugin->getAddUserToCalendarForm($form);
+    $form->addHidden('user_list', '');
+    $form->addButtonSave(get_lang('Add'));
+}
 
 ?>
 <script>
 $(function() {
-<?php
-    // grid definition see the $usergroup>display() function
-    echo Display::grid_js(
-        'usergroups',
-        $url,
-        $columns,
-        $column_model,
-        $extraParams,
-        [],
-        $action_links,
-        true
-    );
-?>
+    <?php
+        // grid definition see the $usergroup>display() function
+        echo Display::grid_js(
+            'usergroups',
+            $url,
+            $columns,
+            $column_model,
+            $extraParams,
+            [],
+            $action_links,
+            true
+        );
+    ?>
     $("#usergroups").jqGrid(
         "navGrid",
         "#usergroups_pager",
-        { edit: false, add: false, del: true, search: false},
+        { edit: false, add: false, del: <?php echo $allowEdit ? 'true' : 'false'; ?>, search: false},
         { height:280, reloadAfterSubmit:false }, // edit options
         { height:280, reloadAfterSubmit:false }, // add options
         { reloadAfterSubmit:false, url: "<?php echo $deleteUrl; ?>" }, // del options
         { width:500 } // search options
     )
+    <?php if ($calendarPlugin) { ?>
     .navButtonAdd('#usergroups_pager',{
-        caption:"<?php echo addslashes($calendarPlugin->get_lang('UpdateCalendar')); ?>",
+        caption:"<?php echo addslashes($calendarPlugin->get_lang('Update calendar')); ?>",
         buttonicon:"ui-icon ui-icon-plus",
         onClickButton: function(a) {
             var userIdList = $("#usergroups").jqGrid('getGridParam', 'selarrrow');
@@ -280,14 +295,16 @@ $(function() {
                 $(".modal-body #add_multiple_calendar_user_list").val(userIdList);
                 $('#myModal').modal();
             } else {
-                alert("<?php echo addslashes(get_lang('SelectStudents')); ?>");
+                alert("<?php echo addslashes(get_lang('Select learners')); ?>");
             }
         },
         position:"last"
-    });
+    })
+    <?php } ?>
+    ;
 });
 </script>
-
+<?php if ($calendarPlugin) { ?>
 <div class="modal fade" id="myModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
     <div class="modal-dialog" role="document">
         <div class="modal-content">
@@ -295,7 +312,7 @@ $(function() {
                 <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                     <span aria-hidden="true">&times;</span></button>
                 <h4 class="modal-title" id="myModalLabel">
-                    <?php echo $calendarPlugin->get_lang('AddMultipleUsersToCalendar'); ?>
+                    <?php echo $calendarPlugin->get_lang('Add multiple users to calendar'); ?>
                 </h4>
             </div>
             <div class="modal-body">
@@ -304,18 +321,18 @@ $(function() {
         </div>
     </div>
 </div>
-
+<?php } ?>
 <?php
 
 $usergroup->showGroupTypeSetting = true;
 // Action handling: Adding a note
-if ($action === 'delete' && is_numeric($_GET['id'])) {
+if ($allowEdit && $action === 'delete' && is_numeric($_GET['id'])) {
     $res = $usergroup->delete_user_rel_group($_GET['user_id'], $_GET['id']);
     Display::addFlash(Display::return_message(get_lang('Deleted'), 'confirmation'));
     header('Location: '.api_get_self().'?id='.$id);
     exit;
-} else {
-    $usergroup->displayToolBarUserGroupUsers();
 }
+
+$usergroup->displayToolBarUserGroupUsers();
 
 Display::display_footer();
